+++ src/com/matriksgeneral/Transpose.java
@@ -1,13 +1,13 @@
 package com.matriksgeneral;
 
 public class Transpose {
-    public static float[][] transpose (float matriks[][], int ord){
+    public static double[][] transpose(double matriks[][], int ord) {
         // mengembalikan transpose matriks
 
-        float temp[][] = matriks;
-        for (int i = 0; i < ord; i++){
-            for (int j = 0; j < ord; j++){
-                temp[i][j] = matriks[j][i];     // baris dan kolom ditukar
+        double temp[][] = matriks;
+        for (int i = 0; i < ord; i++) {
+            for (int j = 0; j < ord; j++) {
+                temp[i][j] = matriks[j][i]; // baris dan kolom ditukar
             }
         }
         return temp;
