+++ src/com/regresi/LinearBerganda.java
@@ -7,7 +7,7 @@ import com.spl.GaussJordan;
 public class LinearBerganda {
     // static Scanner sc = new Scanner(System.in);
 
-    public static void displayMat(float[][] mat) {
+    public static void displayMat(double[][] mat) {
         for (int i = 0; i < mat.length; i++) {
             for (int j = 0; j < mat[i].length; j++) {
                 System.out.print(mat[i][j] + " ");
@@ -17,11 +17,11 @@ public class LinearBerganda {
         }
     }
 
-    public static float[] solveRegresi(float[][] matrix, float[] y) {
+    public static double[] solveRegresi(double[][] matrix, double[] y) {
         // buat matriks augmented [1|matrix|y] : temp
         int row = matrix.length;
         int col = matrix[0].length;
-        float[][] temp = new float[row][col + 2];
+        double[][] temp = new double[row][col + 2];
 
         // Menambahkan angka 1 di kolom pertama sebagai koefisien b0
         for (int i = 0; i < row; i++) {
@@ -40,7 +40,7 @@ public class LinearBerganda {
         }
 
         // Membuat matrix SPL dengan menggunakan Normal Estimation Equation
-        float[][] spl = new float[col + 1][col + 2];
+        double[][] spl = new double[col + 1][col + 2];
         for (int i = 0; i <= col; i++) {
             for (int j = 0; j <= col + 1; j++) {
                 spl[i][j] = 0;
@@ -53,10 +53,10 @@ public class LinearBerganda {
         displayMat(spl);
 
         // Menyelesaikan matrix SPL dengan metode gauss jordan
-        float[][] res = GaussJordan.gaussJordanElimination(spl, col + 1, col + 2);
+        double[][] res = GaussJordan.gaussJordanElimination(spl, col + 1, col + 2);
 
         // Mengambil nilai y saja
-        float[] b = new float[col + 1];
+        double[] b = new double[col + 1];
         for (int i = 0; i <= col; i++) {
             b[i] = res[i][col + 1];
         }
@@ -64,10 +64,10 @@ public class LinearBerganda {
         return b;
     }
 
-    public static void ans(float[][] mat, float y[], int m, int n) {
+    public static void ans(double[][] mat, double y[], int m, int n) {
         // Menyelesaikan regresi linear berganda
-        float[] res = solveRegresi(mat, y);
-        float Result[][] = new float[res.length + 1][2];
+        double[] res = solveRegresi(mat, y);
+        double Result[][] = new double[res.length + 1][2];
         // Mengeluarkan persamaan regresi linear berganda
         System.out.println("\nPersamaan regresi linear berganda yang diperoleh");
         System.out.printf("y = %f", res[0]);
@@ -82,10 +82,10 @@ public class LinearBerganda {
         // Menaksir nilai fungsi
         System.out.println("\n\nMenaksir nilai fungsi");
         System.out.printf("Masukkan %d peubah yang akan ditaksir nilai fungsinya\n", res.length - 1);
-        float[] taksir = new float[res.length];
+        double[] taksir = new double[res.length];
         Scanner sc = new Scanner(System.in);
         for (int i = 0; i < res.length - 1; i++) {
-            taksir[i] = sc.nextFloat();
+            taksir[i] = sc.nextDouble();
         }
 
         for (int i = 0; i < res.length; i++) {
@@ -93,7 +93,7 @@ public class LinearBerganda {
             Result[i][1] = taksir[i];
         }
 
-        float result = res[0];
+        double result = res[0];
         for (int i = 0; i < res.length - 1; i++) {
             result += res[i + 1] * taksir[i];
         }
