+++ src/com/spl/GaussJordan.java
@@ -6,7 +6,7 @@ import java.io.IOException;
 
 // EliminasiGaussJordan Class
 public class GaussJordan {
-    static float[][] mtx = new float[100][100];
+    static double[][] mtx = new double[100][100];
     static int BrsEff;
     static int KolEff;
 
@@ -47,7 +47,7 @@ public class GaussJordan {
     }
 
     private static void swapBrs(int Brs1Idx, int Brs2Idx) {
-        float[] tempBrs = new float[KolEff];
+        double[] tempBrs = new double[KolEff];
         for (int j = 0; j < KolEff; j++) {
             tempBrs[j] = mtx[Brs1Idx][j];
         }
@@ -65,13 +65,13 @@ public class GaussJordan {
         swapBrs(currBrs, pivotBrs);
     }
 
-    private static void bagiBrs(int i, float Bagi) {
+    private static void bagiBrs(int i, double Bagi) {
         for (int j = 0; j < KolEff; j++) {
             mtx[i][j] /= Bagi;
         }
     }
 
-    private static void kurangBrs(int i, int j, float Kali) {
+    private static void kurangBrs(int i, int j, double Kali) {
         for (int k = 0; k < KolEff; k++) {
             mtx[i][k] -= (mtx[j][k] * Kali);
         }
@@ -171,14 +171,14 @@ public class GaussJordan {
             if (currBaseKol > KolEff - 2) {
                 break;
             }
-            float pivotBagi = mtx[currBaseBrs][currBaseKol];
+            double pivotBagi = mtx[currBaseBrs][currBaseKol];
             bagiBrs(currBaseBrs, pivotBagi);
             for (currBrs = currBaseBrs + 1; currBrs < BrsEff; currBrs++) {
-                float multiplier = mtx[currBrs][currBaseKol];
+                double multiplier = mtx[currBrs][currBaseKol];
                 kurangBrs(currBrs, currBaseBrs, multiplier);
             }
             for (currBrs = currBaseBrs - 1; currBrs >= 0; currBrs--) {
-                float multiplier = mtx[currBrs][currBaseKol];
+                double multiplier = mtx[currBrs][currBaseKol];
                 kurangBrs(currBrs, currBaseBrs, multiplier);
             }
             currBaseKol++;
@@ -186,7 +186,7 @@ public class GaussJordan {
         }
     }
 
-    public static float[][] gaussJordanElimination(float[][] m, int M, int N) {
+    public static double[][] gaussJordanElimination(double[][] m, int M, int N) {
         mtx = m;
         BrsEff = M;
         KolEff = N;
@@ -203,14 +203,14 @@ public class GaussJordan {
             if (currBaseKol > KolEff - 2) {
                 break;
             }
-            float pivotBagi = mtx[currBaseBrs][currBaseKol];
+            double pivotBagi = mtx[currBaseBrs][currBaseKol];
             bagiBrs(currBaseBrs, pivotBagi);
             for (currBrs = currBaseBrs + 1; currBrs < BrsEff; currBrs++) {
-                float multiplier = mtx[currBrs][currBaseKol];
+                double multiplier = mtx[currBrs][currBaseKol];
                 kurangBrs(currBrs, currBaseBrs, multiplier);
             }
             for (currBrs = currBaseBrs - 1; currBrs >= 0; currBrs--) {
-                float multiplier = mtx[currBrs][currBaseKol];
+                double multiplier = mtx[currBrs][currBaseKol];
                 kurangBrs(currBrs, currBaseBrs, multiplier);
             }
             currBaseKol++;
@@ -391,7 +391,7 @@ public class GaussJordan {
         }
     }
 
-    public static void ans(float a[][], float b[], int M, int N) {
+    public static void ans(double a[][], double b[], int M, int N) {
         BrsEff = M;
         KolEff = N + 1;
         System.out.println("El");
