+++ src/com/spl/MatriksBalikan.java
@@ -5,10 +5,10 @@ import com.invers.inversMat;
 import java.util.Scanner;
 
 public class MatriksBalikan {
-    public static float[][] matriksbalikan(float[][] A, float[][] B, int ordA) {
+    public static double[][] matriksbalikan(double[][] A, double[][] B, int ordA) {
         // menyelesaikan spl Ax = B, prekondisi A[ordA][ordA], B[ordA][1]
-        float[][] solution = new float[ordA][1];
-        float[][] invA = new float[ordA][ordA];
+        double[][] solution = new double[ordA][1];
+        double[][] invA = new double[ordA][ordA];
 
         invA = inversMat.inversmat(A, ordA);
         solution = Kali.kali(invA, B, ordA, ordA, ordA, 1); // x = A^-1*B
@@ -17,9 +17,9 @@ public class MatriksBalikan {
 
     }
 
-    public static void ans(float[][] A, float[] B, int ordA) {
-        float b[][] = new float[ordA][0];
-        float jwb[][] = matriksbalikan(A, b, ordA);
+    public static void ans(double[][] A, double[] B, int ordA) {
+        double b[][] = new double[ordA][0];
+        double jwb[][] = matriksbalikan(A, b, ordA);
         for (int i = 0; i < ordA; i++) {
             System.out.printf("x[%d] = %f\n", i + 1, jwb[i][0]);
         }
